backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
requests==2.31.0
feedparser==6.0.11
python-dotenv==1.0.1
gevent==24.2.1
gevent-websocket==0.10.1
PyPDF2==3.0.1
python-docx==1.1.0
